feat: parallelized graph writer (#342)

+++ src/deep_neurographs/utils/swc_util.py
@@ -318,9 +318,9 @@ class Reader:
         # Initializations
         content, offset = self.process_content(content)
         swc_dict = {
-            "id": np.zeros((len(content)), dtype=np.int32),
+            "id": np.zeros((len(content)), dtype=int),
             "radius": np.zeros((len(content)), dtype=np.float32),
-            "pid": np.zeros((len(content)), dtype=np.int32),
+            "pid": np.zeros((len(content)), dtype=int),
             "xyz": np.zeros((len(content), 3), dtype=np.float32),
             "soma_nodes": set(),
         }
